requirements.txt
pandas
scikit-learn
gspread
google-auth
kagglehub